GuildsService Messenger handling
