+++ src/Quarrel.ViewModels/Controls/EmojiPickerViewModel.cs
@@ -109,6 +109,7 @@ namespace Quarrel.ViewModels.Controls
                     }
                 }
             });
+
             // TODO: Sort by accuracy
         }
     }
