+++ src/Quarrel.ViewModels/MainViewModel/GuildsMainViewModel.cs
@@ -1,6 +1,8 @@
 ﻿// Copyright (c) Quarrel. All rights reserved.
 
 using DiscordAPI.Models;
+using DiscordAPI.Models.Channels;
+using DiscordAPI.Models.Guilds;
 using GalaSoft.MvvmLight.Command;
 using JetBrains.Annotations;
 using Quarrel.ViewModels.Helpers;
@@ -11,6 +13,8 @@ using Quarrel.ViewModels.Models.Bindables.Channels;
 using Quarrel.ViewModels.Models.Bindables.Guilds;
 using Quarrel.ViewModels.Models.Bindables.Users;
 using Quarrel.ViewModels.Models.Interfaces;
+using Quarrel.ViewModels.ViewModels.Messages.Gateway;
+using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
@@ -37,10 +41,7 @@ namespace Quarrel.ViewModels
             {
                 if (guildListItem is BindableGuild bGuild)
                 {
-                    Task.Run(() =>
-                    {
-                        MessengerInstance.Send(new GuildNavigateMessage(bGuild));
-                    });
+                    CurrentGuild = bGuild;
                 }
                 else if (guildListItem is BindableGuildFolder bindableGuildFolder)
                 {
@@ -78,7 +79,13 @@ namespace Quarrel.ViewModels
                     _currentGuild.Selected = false;
                 }
 
-                Set(ref _currentGuild, value);
+                if (Set(ref _currentGuild, value))
+                {
+                    Task.Run(() =>
+                    {
+                        MessengerInstance.Send(new GuildNavigateMessage(_currentGuild));
+                    });
+                }
 
                 if (_currentGuild != null)
                 {
@@ -116,63 +123,233 @@ namespace Quarrel.ViewModels
 
         private void RegisterGuildsMessages()
         {
-            MessengerInstance.Register<GuildNavigateMessage>(this, m =>
+            MessengerInstance.Register<GatewayReadyMessage>(this, m =>
             {
-                if (CurrentGuild != m.Guild)
+                _dispatcherHelper.CheckBeginInvokeOnUi(() =>
                 {
-                    BindableChannel channel =
-                        m.Guild.Channels.FirstOrDefault(x => x.IsTextChannel && x.Permissions.ReadMessages);
-                    BindableGuildMember currentGuildMember;
-
-                    if (!m.Guild.IsDM)
+                    IDictionary<string, ReadState> readStates = new ConcurrentDictionary<string, ReadState>();
+                    foreach (var state in m.EventData.ReadStates)
                     {
-                        currentGuildMember = _guildsService.GetGuildMember(_currentUserService.CurrentUser.Model.Id, m.Guild.Model.Id);
-                    }
-                    else
-                    {
-                        currentGuildMember = new BindableGuildMember(
-                            new DiscordAPI.Models.GuildMember()
-                            {
-                                User = _currentUserService.CurrentUser.Model,
-                            },
-                            "DM",
-                            _currentUserService.CurrentUser.Presence);
+                        readStates.Add(state.Id, state);
                     }
 
-                    _dispatcherHelper.CheckBeginInvokeOnUi(() =>
-                    {
-                        CurrentChannel = channel;
-                        CurrentGuild = m.Guild;
-                        BindableMessages.Clear();
+                    // Add DM
+                    var dmGuild = new BindableGuild(new Guild() { Name = "DM", Id = "DM" }) { Position = -1 };
 
-                        if (m.Guild.IsDM)
+                    _guildsService.AddOrUpdateGuild(dmGuild.Model.Id, dmGuild);
+
+                    // Add DM channels
+                    if (m.EventData.PrivateChannels != null && m.EventData.PrivateChannels.Any())
+                    {
+                        foreach (var channel in m.EventData.PrivateChannels)
                         {
-                            CurrentBindableMembers.Clear();
+                            BindableChannel bChannel = new BindableChannel(channel);
+
+                            ChannelOverride cSettings = _channelsService.GetChannelSettings(channel.Id);
+                            if (cSettings != null)
+                            {
+                                bChannel.Muted = cSettings.Muted;
+                            }
+
+                            dmGuild.Channels.Add(bChannel);
+
+                            if (readStates.ContainsKey(bChannel.Model.Id))
+                            {
+                                bChannel.ReadState = readStates[bChannel.Model.Id];
+                                if (bChannel.ReadState.LastMessageId != bChannel.Model.LastMessageId)
+                                {
+                                    bChannel.ReadState.MentionCount++;
+                                }
+                            }
+
+                            _channelsService.AddOrUpdateChannel(bChannel.Model.Id, bChannel);
                         }
 
-                        CurrentGuildMember = currentGuildMember;
-                    });
-
-                    if (channel != null)
-                    {
-                        MessengerInstance.Send(new ChannelNavigateMessage(channel));
+                        // Sort by last message timestamp
+                        dmGuild.Channels = new ObservableCollection<BindableChannel>(dmGuild.Channels.OrderByDescending(x => Convert.ToUInt64(x.Model.LastMessageId)).ToList());
                     }
 
-                    _analyticsService.Log(
-                        m.Guild.IsDM ?
-                        Constants.Analytics.Events.OpenDMs :
-                        Constants.Analytics.Events.OpenGuild,
-                        ("guild-id", m.Guild.Model.Id));
-                }
-            });
+                    foreach (var guild in m.EventData.Guilds)
+                    {
+                        BindableGuild bGuild = new BindableGuild(guild);
 
+                        // Handle guild settings
+                        GuildSetting gSettings = _guildsService.GetGuildSetting(guild.Id);
+                        if (gSettings != null)
+                        {
+                            bGuild.IsMuted = gSettings.Muted;
+                        }
+
+                        // Guild Order
+                        bGuild.Position = m.EventData.Settings.GuildOrder.IndexOf(x => x == bGuild.Model.Id);
+
+                        if (guild.Unavailable)
+                        {
+                            // TODO:
+                            // This should be updated to display information about unavailable guild
+                            // We also need a mechanism to readd the guild once it becomes available
+                            continue;
+                        }
+
+                        // This is needed to fix ordering when multiple categories have the same position
+                        var categories = guild.Channels.Where(x => x.Type == 4).ToList();
+
+                        foreach (var group in categories.GroupBy(x => x.Position).OrderBy(x => x.Key))
+                        {
+                            foreach (var channel in group.Skip(1))
+                            {
+                                bool shouldDo = false;
+                                foreach (var category in categories)
+                                {
+                                    if (category.Id == channel.Id)
+                                    {
+                                        shouldDo = true;
+                                    }
+
+                                    if (shouldDo)
+                                    {
+                                        category.Position += 1;
+                                    }
+                                }
+                            }
+                        }
+
+                        if (guild.VoiceStates != null)
+                        {
+                            foreach (var state in guild.VoiceStates)
+                            {
+                                _voiceService.VoiceStates[state.UserId] = state;
+                            }
+                        }
+
+                        // Guild Channels
+                        foreach (var channel in guild.Channels)
+                        {
+                            channel.GuildId = guild.Id;
+                            IEnumerable<VoiceState> state = guild.VoiceStates?.Where(x => x.ChannelId == channel.Id);
+                            BindableChannel bChannel = new BindableChannel(channel, state);
+
+                            // Handle channel settings
+                            ChannelOverride cSettings = _channelsService.GetChannelSettings(channel.Id);
+                            if (cSettings != null)
+                            {
+                                bChannel.Muted = cSettings.Muted;
+                            }
+
+                            // Find parent position
+                            if (!string.IsNullOrEmpty(bChannel.ParentId) && bChannel.ParentId != bChannel.Model.Id)
+                            {
+                                bChannel.ParentPostion = guild.Channels.First(x => x.Id == bChannel.ParentId).Position;
+                            }
+                            else
+                            {
+                                bChannel.ParentPostion = -1;
+                            }
+
+                            if (readStates.ContainsKey(bChannel.Model.Id))
+                            {
+                                bChannel.ReadState = readStates[bChannel.Model.Id];
+                            }
+
+                            bGuild.Channels.Add(bChannel);
+                            _channelsService.AddOrUpdateChannel(bChannel.Model.Id, bChannel);
+                        }
+
+                        bGuild.Channels = new ObservableCollection<BindableChannel>(bGuild.Channels.OrderBy(x => x.AbsolutePostion).ToList());
+
+                        bGuild.Model.Channels = null;
+
+                        _guildsService.GuildUsers.AddOrUpdate(guild.Id, new ConcurrentDictionary<string, BindableGuildMember>());
+                        foreach (var user in guild.Members)
+                        {
+                            BindableGuildMember bgMember = new BindableGuildMember(user, guild.Id);
+                            _guildsService.GuildUsers[guild.Id].TryAdd(bgMember.Model.User.Id, bgMember);
+                        }
+
+                        // Guild Roles
+                        foreach (var role in guild.Roles)
+                        {
+                            _cacheService.Runtime.SetValue(Constants.Cache.Keys.GuildRole, role, guild.Id + role.Id);
+                        }
+
+                        // Guild Presences
+                        foreach (var presence in guild.Presences)
+                        {
+                            MessengerInstance.Send(new GatewayPresenceUpdatedMessage(presence.User.Id, presence));
+                        }
+
+                        _guildsService.AddOrUpdateGuild(bGuild.Model.Id, bGuild);
+                    }
+
+                    BindableGuilds.Clear();
+                    BindableGuilds.Add(_guildsService.GetGuild("DM"));
+                    foreach (var folder in m.EventData.Settings.GuildFolders)
+                    {
+                        BindableGuildFolder bindableFolder = new BindableGuildFolder(folder) { IsCollapsed = folder.GuildIds.Count() > 1 };
+                        BindableGuilds.Add(bindableFolder);
+                        foreach (var guildId in folder.GuildIds)
+                        {
+                            BindableGuild guild = _guildsService.GetGuild(guildId);
+                            guild.FolderId = folder.Id;
+                            guild.IsCollapsed = bindableFolder.IsCollapsed;
+                            BindableGuilds.Add(guild);
+                        }
+                    }
+
+                    CurrentGuild = dmGuild;
+                });
+            });
+            MessengerInstance.Register<GuildNavigateMessage>(this, m =>
+            {
+                BindableChannel channel =
+                m.Guild.Channels.FirstOrDefault(x => x.IsTextChannel && x.Permissions.ReadMessages);
+                BindableGuildMember currentGuildMember;
+
+                if (!m.Guild.IsDM)
+                {
+                    currentGuildMember = _guildsService.GetGuildMember(_currentUserService.CurrentUser.Model.Id, m.Guild.Model.Id);
+                }
+                else
+                {
+                    currentGuildMember = new BindableGuildMember(
+                        new GuildMember()
+                        {
+                            User = _currentUserService.CurrentUser.Model,
+                        },
+                        "DM",
+                        _currentUserService.CurrentUser.Presence);
+                }
+
+                _dispatcherHelper.CheckBeginInvokeOnUi(() =>
+                {
+                    CurrentChannel = channel;
+                    BindableMessages.Clear();
+
+                    if (m.Guild.IsDM)
+                    {
+                        CurrentBindableMembers.Clear();
+                    }
+
+                    CurrentGuildMember = currentGuildMember;
+                });
+
+                if (channel != null)
+                {
+                    MessengerInstance.Send(new ChannelNavigateMessage(channel));
+                }
+
+                _analyticsService.Log(
+                    m.Guild.IsDM ?
+                    Constants.Analytics.Events.OpenDMs :
+                    Constants.Analytics.Events.OpenGuild,
+                    ("guild-id", m.Guild.Model.Id));
+            });
             MessengerInstance.Register<GatewayGuildCreatedMessage>(this, m =>
             {
                 BindableGuild guild = new BindableGuild(m.Guild);
                 _guildsService.AddOrUpdateGuild(m.Guild.Id, guild);
                 _dispatcherHelper.CheckBeginInvokeOnUi(() => { BindableGuilds.Insert(1, guild); });
             });
-
             MessengerInstance.Register<GatewayGuildDeletedMessage>(this, m =>
             {
                 BindableGuild guild;
@@ -183,7 +360,95 @@ namespace Quarrel.ViewModels
                     _dispatcherHelper.CheckBeginInvokeOnUi(() => { BindableGuilds.Remove(guild); });
                 }
             });
+            MessengerInstance.Register<GatewayChannelCreatedMessage>(this, m =>
+            {
+                string guildId = "DM";
+                if (m.Channel is GuildChannel gChannel)
+                {
+                    guildId = gChannel.GuildId;
+                }
 
+                var bChannel = new BindableChannel(m.Channel);
+                if (bChannel.Model.Type != 4 && bChannel.ParentId != null)
+                {
+                    BindableChannel parentChannel = _channelsService.GetChannel(bChannel.ParentId);
+                    bChannel.ParentPostion = parentChannel != null ? parentChannel.Position : 0;
+                }
+                else if (bChannel.ParentId == null)
+                {
+                    bChannel.ParentPostion = -1;
+                }
+
+                BindableGuild guild = _guildsService.GetGuild(guildId);
+                if (guild != null)
+                {
+                    for (int i = 0; i < guild.Channels.Count; i++)
+                    {
+                        if (guild.Channels[i].AbsolutePostion > bChannel.AbsolutePostion)
+                        {
+                            _dispatcherHelper.CheckBeginInvokeOnUi(() =>
+                            {
+                                guild.Channels.Insert(i, bChannel);
+                            });
+                            break;
+                        }
+                    }
+                }
+
+                _channelsService.AddOrUpdateChannel(bChannel.Model.Id, bChannel);
+            });
+            MessengerInstance.Register<GatewayChannelDeletedMessage>(this, m =>
+            {
+                _dispatcherHelper.CheckBeginInvokeOnUi(() =>
+                {
+                    BindableChannel currentChannel = _channelsService.GetChannel(m.Channel.Id);
+                    if (currentChannel != null)
+                    {
+                        BindableGuild guild = _guildsService.GetGuild(currentChannel.GuildId);
+                        if (guild != null)
+                        {
+                            guild.Channels.Remove(currentChannel);
+                        }
+
+                        _channelsService.RemoveChannel(m.Channel.Id);
+                    }
+                });
+            });
+            MessengerInstance.Register<GatewayGuildChannelUpdatedMessage>(this, m =>
+            {
+                _dispatcherHelper.CheckBeginInvokeOnUi(() =>
+                {
+                    var bChannel = _channelsService.GetChannel(m.Channel.Id ?? "DM");
+                    bChannel.Model = m.Channel;
+
+                    if (bChannel.Model.Type != 4 && bChannel.ParentId != null)
+                    {
+                        BindableChannel parentChannel = _channelsService.GetChannel(bChannel.ParentId);
+                        bChannel.ParentPostion = parentChannel != null ? parentChannel.Position : 0;
+                    }
+                    else if (bChannel.ParentId == null)
+                    {
+                        bChannel.ParentPostion = -1;
+                    }
+
+                    BindableGuild guild = _guildsService.GetGuild(m.Channel.GuildId);
+                    if (guild != null)
+                    {
+                        guild.Channels.Remove(bChannel);
+                        for (int i = 0; i < guild.Channels.Count; i++)
+                        {
+                            if (guild.Channels[i].AbsolutePostion > bChannel.AbsolutePostion)
+                            {
+                                _dispatcherHelper.CheckBeginInvokeOnUi(() =>
+                                {
+                                    guild.Channels.Insert(i, bChannel);
+                                });
+                                break;
+                            }
+                        }
+                    }
+                });
+            });
             MessengerInstance.Register<GatewayUserSettingsUpdatedMessage>(this, m =>
             {
                 _dispatcherHelper.CheckBeginInvokeOnUi(() =>
@@ -193,6 +458,22 @@ namespace Quarrel.ViewModels
                         // TODO: Handle guild reorder.
                     }
                 });
+            });
+            MessengerInstance.Register<GatewayGuildMembersChunkMessage>(this, m =>
+            {
+                _guildsService.GuildUsers.TryGetValue(m.GuildMembersChunk.GuildId, out var guild);
+                foreach (var member in m.GuildMembersChunk.Members)
+                {
+                    guild.TryAdd(member.User.Id, new BindableGuildMember(member, m.GuildMembersChunk.GuildId));
+                }
+
+                if (m.GuildMembersChunk.Presences != null)
+                {
+                    foreach (var presence in m.GuildMembersChunk.Presences)
+                    {
+                        _presenceService.UpdateUserPrecense(presence.User.Id, presence);
+                    }
+                }
             });
         }
     }
